Pass 3 Task 1: extract ghost data types to GhostTypes.cs
Move 21 data types (12 classes, 7 structs, 2 enums) from top of
GhostVisualBuilder.cs to new GhostTypes.cs. Types are already
top-level in namespace Parsek — no reference updates needed.
GhostVisualBuilder.cs -195 lines. Zero logic changes.
